Fix writes to closed channels
This fixes a timing condition where the watch loop of a signal could
still be waiting to write an event when the channel is closed. This
condition resulted in a panic due to the closed channel write. Now, the
channels are never closed but the goroutines are guaranteed to exit once
the signals are stopped by selecting all writes along with reads of the
closer channel.