Replace embedded FS watcher with signal providers
The trigger CDC engine implementation previously embedded the FS watcher
logic. This is factored out and replaced with customizable signals. The
default signal is now set to a combination of the FS watcher and a timer
after discovering that some resource constrained or containerized
environments don't always emit FS events consistently. I didn't get to
an exact root cause but the tests would intermittently pass or fail on
some, but not all, machines.