Add bootstrap events and more code docs
The bootstrap events are useful for systems that accumulate state based
on the CDC stream and then later encounter a critical fault. The
standard recovery is to bootstrap but bootstrap records are otherwise
identical to INSERT change records. This leaves the consuming system
with no way of knowing whether any currently persisted state is invalid
such as a record that was deleted but the system failed to correctly
process the delete. In this case, a bootstrap event can be a signal to
purge any persisted state relevant to a table before handling the
bootstrap events.